panel product back reload

+++ app/panel/views.py
@@ -545,41 +545,43 @@ def products(request):
     manager = Shop.objects.values_list('customuser_id', flat=True).distinct()
     shops=Shop.objects.values_list('customuser_id','slug' ).distinct()
     address = str([i for i in str(request.path).split('/') if i][0])
-    for u in users:
-        for m in manager:
-            if u == m and request.user.id == u:
-                for s,slug in shops:
-                    if s==m:
-                        name=eval(slug)
-                        products = name.objects.all()
-                        paginator = Paginator(products, 50)
-                        page_number = request.GET.get('page')
-                        page_obj = paginator.get_page(page_number)
-                        if request.method == 'POST':
-                            check_ = request.POST.getlist("check_")
-                            checkbool = request.POST.get("checkbool")
-                            query_name = request.POST.get('name')
-                            if query_name:
-                                product = name.objects.filter(Q(name__icontains=query_name)).order_by('name')
-                                paginator = Paginator(product, 50)
-                                page_number = request.GET.get('page')
-                                page_obj = paginator.get_page(page_number)
+    if request.user.is_authenticated:
+        for u in users:
+            for m in manager:
+                if u == m and request.user.id == u:
+                    for s,slug in shops:
+                        if s==m:
+                            name=eval(slug)
+                            products = name.objects.all()
+                            paginator = Paginator(products, 50)
+                            page_number = request.GET.get('page')
+                            page_obj = paginator.get_page(page_number)
+                            if request.method == 'POST':
+                                check_ = request.POST.getlist("check_")
+                                checkbool = request.POST.get("checkbool")
+                                query_name = request.POST.get('name')
+                                if query_name:
+                                    product = name.objects.filter(Q(name__icontains=query_name)).order_by('name')
+                                    paginator = Paginator(product, 50)
+                                    page_number = request.GET.get('page')
+                                    page_obj = paginator.get_page(page_number)
+                                    return render(request, 'panel/products.html', {'address':address,'page_obj': page_obj, 'products': products})
+                                item = [i.split(',') for i in check_][0]
+                                for i in item:
+                                    if i == 'on':
+                                        item.pop(0)
+                                items = list(map(int, item))
+                                if checkbool and checkbool != 'delete':
+                                    name.objects.filter(pk__in=items).update(status=checkbool)
+                                    return render(request, 'panel/products.html', {'address':address,'page_obj':page_obj,'products': products})
+                                if checkbool == 'delete':
+                                    name.objects.filter(pk__in=items).delete()
+                                    return render(request, 'panel/products.html', {'address': address, 'page_obj': page_obj, 'products': products})
+                            else:
                                 return render(request, 'panel/products.html', {'address':address,'page_obj': page_obj, 'products': products})
-                            item = [i.split(',') for i in check_][0]
-                            for i in item:
-                                if i == 'on':
-                                    item.pop(0)
-                            items = list(map(int, item))
-                            if checkbool and checkbool != 'delete':
-                                name.objects.filter(pk__in=items).update(status=checkbool)
-                                return render(request, 'panel/products.html', {'address':address,'page_obj':page_obj,'products': products})
-                            if checkbool == 'delete':
-                                name.objects.filter(pk__in=items).delete()
-                                return render(request, 'panel/products.html', {'address': address, 'page_obj': page_obj, 'products': products})
-                        else:
-                            return render(request, 'panel/products.html', {'address':address,'page_obj': page_obj, 'products': products})
-                        return render(request, 'panel/products.html', {'address':address,'page_obj':page_obj,'products': products})
-
+                            return render(request, 'panel/products.html', {'address':address,'page_obj':page_obj,'products': products})
+    else:
+        return redirect ('/login')
 
 
 #Просмотр продукта
